// frontend/src/types/coupon.ts
export interface RewardItem {
  item: string;
  amount: number;
}

export interface Coupon {
  id: number;
  gameId: number;
  korName: string;
  engName: string;
  code: string;
  description?: string;
  server: string;
  rewards: RewardItem[];
  createdAt: string;
  expiredAt?: string;
  slug: string;
  quickUrl?: string;
}